requirements.txt
pandas
matplotlib
numpy
requests
requests-cache
tabulate
python-dotenv
Pillow
scikit-learn
os
sys
dotenv
cachetools
json
paho-mqtt
ssl
base64

